CSV Support (#41)
* first draft of C++ code for reading CSV

* prepend row number in test output

* gen simple csv data file

* first csv test

* handle pd.read_csv in hiframes

* CsvReader IR node

* more CSV reader node handling

* fix csv output array type

* csv reader lowering handling

* using carray of MemInfo*

* using meminfo, first sketch for lowering

* add numpy include dir

* fix _csv_read intrinsic (return codegen)

* fix return type of _csv_read

* fix csv node file_name

* fix const tuple to str cast

* update numba to c type map in CSV

* fix input types of _csv_read intrinsic signature (const not allowed)

* fix _csv_read codegen syntax errors

* fix csv reader llvm types

* fix type of csv dtypes

* create tuple of arrays to return in _csv_read codegen

* use hpat CTypes in csv read

* some experiments in comments, experimental breakpoint

* Revert "some experiments in comments, experimental breakpoint"

This reverts commit 7823e9106b63040e84835b70897490042b617da0.

* using HPAT types, printing table after read

* removing dead/unneeded code

* proper checks in importing meminfo func

* fixes in meminfo func calls (needs Numba #3272)

* fixing distributed reading and compiler warnings

* fix single column case in CSV reader backend

* fixing skipping incomplete lines

* lower level stuff for datetime support

* parse CSV date test

* refactor read_csv dtype arg parsing, support str

* read_csv frontend support parse_dates

* starting with going through pandas, removed warnings

* moved extern decl to header

* fixing decls

* trying to convert df to tuple, incomplete

* fix df unbox signature

* chunking and lazy file-IO for file-like IO object

* streamlining chunk boundary exchange

* provide wrappers for getting chunk reader from file/string; removing old csv_read functionality from C

* reducing warnings with GNU compilers

* using HPAT's formmula for 1D_Blocking for csv chunking

* updated comments/doc

* WIP: try generating objmode function, hits globals bug

* fix csv reader pandas wrapper generation

* support datetime64 parsing in csv reader pandas wrapper

* support strings in csv reader pandas wrapper

* avoid conflicts: remove datetime in utils, add regex back in setup

* fix CSV reduce op

* _csv.cpp pass pointer to allreduce

* CSV reader: use is_parallel flag

* csv dist analysis fix

* csv parallel test

* csv reader fix dist array size/start/counts

* fixing csv chunking

* set read_csv output type to series instead of array

* csv remove dead support

* fix for CSV parallel crash

* fixing chunking of streams

* move csv stuff to _csv.cpp

* cleanup leftover print, etc

* fix csv dt64 type to series

+++ hpat/_str_ext.cpp
@@ -347,7 +347,7 @@ void set_string_array_range(uint32_t *out_offsets, char *out_data,
         curr_offset = out_offsets[start_str_ind];
 
     // set offsets
-    for (size_t i=0; i<num_strs; i++)
+    for (size_t i=0; i<(size_t)num_strs; i++)
     {
         out_offsets[start_str_ind+i] = curr_offset;
         int32_t len = in_offsets[i+1]-in_offsets[i];
